requirements.txt
fastapi==0.112.2             
uvicorn==0.30.1              
Pillow==8.4.0                
requests==2.31.0             
python-dotenv==1.0.1         
google-cloud-logging==3.11.2 
pytest==8.3.2                
